fix FuelXLap type to float32

+++ type_graphic.go
@@ -42,7 +42,7 @@ type AccGraphic struct {
 	TCCut                    int32
 	EngineMap                int32
 	ABS                      int32
-	FuelXLap                 int32
+	FuelXLap                 float32
 	RainLights               int32
 	FlashingLights           int32
 	LightsStage              int32
